backends/stable_diffusion_tf/requirements.txt
tensorflow==2.10.0
h5py==3.7.0
Pillow==9.2.0
tqdm==4.64.1
ftfy==6.1.1
regex==2022.9.13